clk: mediatek: Add new pll type LC used in mt6589
Mediateks mt6589 SoC has a different type of PLL which is used for some clks.
It is called LC.
Differences to the already implemented type are:
- post divider ratio is two bits instead of three
- no power register exists
- some plls have a fixed rate

This patch adds a pd_mask to the mtk_pll_data struct and adds flags for the
PLLs without a power register and for those which have a fixed rate and
a post divider.

Signed-off-by: Matthias Brugger <matthias.bgg@kernel.org>